feat: Update structure project so it is more easily expandable
